refactor test_utils.py (to match proper source)
