+++ bevroad/test_utils.py
@@ -1,0 +1,157 @@
+import math
+import os
+import tempfile
+import unittest
+
+from bevroad.utils import (
+    global_to_local,
+    load_calibration,
+    load_config,
+    parse_csv,
+    save_config,
+    transform_point,
+)
+
+
+class TestTransformAndCoordinateMath(unittest.TestCase):
+    def test_transform_point_identity(self):
+        # No rotation, origin
+        gx, gy = transform_point(1.0, 2.0, 0.0, 0.0, 0.0)
+        self.assertAlmostEqual(gx, 1.0)
+        self.assertAlmostEqual(gy, 2.0)
+
+    def test_transform_point_rotation_and_translation(self):
+        # Robot at (10.0, 5.0) facing 90 degrees (pi/2 radians, pointing in +Y direction)
+        # Local forward point x_local=2.0, y_local=0.0 should end up at global X=10.0, Y=7.0
+        gx, gy = transform_point(2.0, 0.0, 10.0, 5.0, math.pi / 2.0)
+        self.assertAlmostEqual(gx, 10.0)
+        self.assertAlmostEqual(gy, 7.0)
+
+    def test_transform_point_origin_no_rotation(self):
+        # Robot at (0, 0) facing 0 radians
+        gx, gy = transform_point(x_local=2.0, y_local=1.0, robot_x=0.0, robot_y=0.0, heading=0.0)
+        self.assertAlmostEqual(gx, 2.0)
+        self.assertAlmostEqual(gy, 1.0)
+
+    def test_transform_point_with_translation_and_rotation(self):
+        # Robot at (5.0, 10.0) facing pi/2 (90 degrees, pointing along +Y)
+        # local forward point X_local=3.0, Y_local=0.0 should be at global X = 5.0, Y = 13.0
+        gx, gy = transform_point(x_local=3.0, y_local=0.0, robot_x=5.0, robot_y=10.0, heading=math.pi / 2.0)
+        self.assertAlmostEqual(gx, 5.0)
+        self.assertAlmostEqual(gy, 13.0)
+
+        # local left point X_local=0.0, Y_local=2.0 (left of robot) should be at global X=3.0, Y=10.0 (pointing west)
+        gx, gy = transform_point(x_local=0.0, y_local=2.0, robot_x=5.0, robot_y=10.0, heading=math.pi / 2.0)
+        self.assertAlmostEqual(gx, 3.0)
+        self.assertAlmostEqual(gy, 10.0)
+
+    def test_global_to_local_inverse(self):
+        # Test various positions and orientations to ensure global_to_local is the true mathematical inverse
+        test_cases = [
+            (1.0, 2.0, 0.0, 0.0, 0.0),
+            (2.0, -1.0, 10.0, 5.0, math.pi / 2.0),
+            (-3.5, 4.2, -20.0, 35.0, -math.pi / 4.0),
+            (0.0, 0.0, 100.0, -50.0, 2.5),
+        ]
+        for xl, yl, rx, ry, heading in test_cases:
+            gx, gy = transform_point(xl, yl, rx, ry, heading)
+            xl_inv, yl_inv = global_to_local(gx, gy, rx, ry, heading)
+            self.assertAlmostEqual(xl_inv, xl)
+            self.assertAlmostEqual(yl_inv, yl)
+
+
+class TestCsvParsing(unittest.TestCase):
+    def test_parse_csv_valid_data(self):
+        with tempfile.TemporaryDirectory() as tmpdir:
+            csv_path = os.path.join(tmpdir, 'overview.csv')
+            with open(csv_path, 'w', encoding='utf-8') as f:
+                f.write('img-0000.jpg,12.5,29.86,1.50,-0.0356\n')
+                f.write('img-0001.jpg,13.0,30.36,1.48,-0.0567\n')
+
+            records = parse_csv(csv_path)
+            self.assertEqual(len(records), 2)
+            self.assertEqual(records[0]['filename'], 'img-0000.jpg')
+            self.assertEqual(records[0]['ts'], 12.5)
+            self.assertEqual(records[0]['x'], 29.86)
+            self.assertEqual(records[0]['y'], 1.50)
+            self.assertAlmostEqual(records[0]['heading'], -0.0356)
+
+    def test_parse_csv_valid_and_invalid(self):
+        with tempfile.TemporaryDirectory() as tmpdir:
+            csv_path = os.path.join(tmpdir, 'overview.csv')
+            with open(csv_path, 'w', encoding='utf-8') as f:
+                f.write('img-0000.jpg,1.5,10.0,20.0,1.570796\n')
+                f.write('invalid_line,1.2\n')  # Too short
+                f.write('img-0001.jpg,abc,30.0,40.0,0.0\n')  # Invalid float
+                f.write('img-0002.jpg,3.5,30.0,40.0,0.0\n')
+
+            records = parse_csv(csv_path)
+            self.assertEqual(len(records), 2)
+            self.assertEqual(records[0]['filename'], 'img-0000.jpg')
+            self.assertEqual(records[0]['ts'], 1.5)
+            self.assertEqual(records[0]['x'], 10.0)
+            self.assertEqual(records[0]['y'], 20.0)
+            self.assertAlmostEqual(records[0]['heading'], 1.570796)
+
+            self.assertEqual(records[1]['filename'], 'img-0002.jpg')
+            self.assertEqual(records[1]['ts'], 3.5)
+            self.assertEqual(records[1]['x'], 30.0)
+            self.assertEqual(records[1]['y'], 40.0)
+            self.assertAlmostEqual(records[1]['heading'], 0.0)
+
+
+class TestJsonConfig(unittest.TestCase):
+    def setUp(self):
+        self.test_dir = tempfile.TemporaryDirectory()
+
+    def tearDown(self):
+        self.test_dir.cleanup()
+
+    def test_load_config_non_existent(self):
+        non_existent_path = os.path.join(self.test_dir.name, 'does_not_exist.json')
+        config = load_config(non_existent_path)
+        self.assertEqual(config, {})
+
+    def test_save_and_load_config(self):
+        config_path = os.path.join(self.test_dir.name, 'test_config.json')
+        test_data = {
+            'top_y': 100,
+            'bottom_y': 200,
+            'top_width': 300,
+            'bottom_width': 400,
+            'offset_x': 50,
+            'show_grid': 1,
+        }
+
+        save_config(config_path, test_data)
+
+        # Verify file was written
+        self.assertTrue(os.path.exists(config_path))
+
+        # Load and verify content
+        loaded_data = load_config(config_path)
+        self.assertEqual(loaded_data, test_data)
+
+    def test_load_config_invalid_json(self):
+        invalid_path = os.path.join(self.test_dir.name, 'invalid.json')
+        with open(invalid_path, 'w') as f:
+            f.write('{invalid-json:')
+
+        config = load_config(invalid_path)
+        self.assertEqual(config, {})
+
+    def test_load_calibration_non_existent(self):
+        non_existent_path = os.path.join(self.test_dir.name, 'does_not_exist.json')
+        with self.assertRaises(SystemExit):
+            load_calibration(non_existent_path)
+
+    def test_load_calibration_invalid_json(self):
+        invalid_path = os.path.join(self.test_dir.name, 'invalid.json')
+        with open(invalid_path, 'w') as f:
+            f.write('{invalid-json:')
+        with self.assertRaises(SystemExit):
+            load_calibration(invalid_path)
+
+
+if __name__ == '__main__':
+    unittest.main()
